move VoterData to voter_data mod
